pip_requirements.txt
annotated-types==0.7.0
awkward==1.10.5
awkward0==0.15.5
bokeh==3.4.1
boost_histogram==1.4.1
cachetools==5.3.3
click==8.1.7
cloudpickle==3.0.0
coffea==0.7.21
contourpy==1.2.1
correctionlib==2.2.2
cycler==0.12.1
dask==2024.5.2
dask-jobqueue==0.8.5
distributed==2024.5.2
fonttools==4.53.0
fsspec==2024.6.0
h5py==3.11.0
hdf5plugin==4.4.0
hist==2.7.3
histoprint==2.4.0
hjson==3.1.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
Jinja2==3.1.4
kiwisolver==1.4.5
llvmlite==0.42.0
locket==1.0.0
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mplhep==0.3.49
mplhep-data==0.0.3
msgpack==1.0.8
numba==0.59.1
numpy==1.23.5
packaging==24.1
pandas==2.2.2
partd==1.4.2
pillow==10.3.0
psutil==5.9.8
pyarrow==16.1.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
scipy==1.13.1
six==1.16.0
sortedcontainers==2.4.0
tblib==3.0.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
uhi==0.4.0
uproot==4.3.7
uproot3==3.14.4
uproot3-methods==0.10.1
urllib3==1.26.18
xrootd==5.6.9
xyzservices==2024.6.0
zict==3.0.0
zipp==3.19.2
